# statistics_manager_service/models/co2_forecast.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from statistics_manager_service.models.base_model import Model
from statistics_manager_service.models.co2_units import CO2Units
from statistics_manager_service import util

from statistics_manager_service.models.co2_units import CO2Units  # noqa: E501

class CO2Forecast(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, datetime=None, request=None, value=None, unit=None, quality=None, updated_at=None):  # noqa: E501
        """CO2Forecast - a model defined in OpenAPI

        :param datetime: The datetime of this CO2Forecast.  # noqa: E501
        :type datetime: datetime
        :param request: The request of this CO2Forecast.  # noqa: E501
        :type request: datetime
        :param value: The value of this CO2Forecast.  # noqa: E501
        :type value: float
        :param unit: The unit of this CO2Forecast.  # noqa: E501
        :type unit: CO2Units
        :param quality: The quality of this CO2Forecast.  # noqa: E501
        :type quality: float
        :param updated_at: The updated_at of this CO2Forecast.  # noqa: E501
        :type updated_at: datetime
        """
        self.openapi_types = {
            'datetime': datetime,
            'request': datetime,
            'value': float,
            'unit': CO2Units,
            'quality': float,
            'updated_at': datetime
        }

        self.attribute_map = {
            'datetime': 'datetime',
            'request': 'request',
            'value': 'value',
            'unit': 'unit',
            'quality': 'quality',
            'updated_at': 'updated_at'
        }

        self._datetime = datetime
        self._request = request
        self._value = value
        self._unit = unit
        self._quality = quality
        self._updated_at = updated_at

    @classmethod
    def from_dict(cls, dikt) -> 'CO2Forecast':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The CO2Forecast of this CO2Forecast.  # noqa: E501
        :rtype: CO2Forecast
        """
        return util.deserialize_model(dikt, cls)

    @property
    def datetime(self) -> datetime:
        """Gets the datetime of this CO2Forecast.


        :return: The datetime of this CO2Forecast.
        :rtype: datetime
        """
        return self._datetime

    @datetime.setter
    def datetime(self, datetime: datetime):
        """Sets the datetime of this CO2Forecast.


        :param datetime: The datetime of this CO2Forecast.
        :type datetime: datetime
        """
        if datetime is None:
            raise ValueError("Invalid value for `datetime`, must not be `None`")  # noqa: E501

        self._datetime = datetime

    @property
    def request(self) -> datetime:
        """Gets the request of this CO2Forecast.


        :return: The request of this CO2Forecast.
        :rtype: datetime
        """
        return self._request

    @request.setter
    def request(self, request: datetime):
        """Sets the request of this CO2Forecast.


        :param request: The request of this CO2Forecast.
        :type request: datetime
        """
        if request is None:
            raise ValueError("Invalid value for `request`, must not be `None`")  # noqa: E501

        self._request = request

    @property
    def value(self) -> float:
        """Gets the value of this CO2Forecast.


        :return: The value of this CO2Forecast.
        :rtype: float
        """
        return self._value

    @value.setter
    def value(self, value: float):
        """Sets the value of this CO2Forecast.


        :param value: The value of this CO2Forecast.
        :type value: float
        """
        if value is None:
            raise ValueError("Invalid value for `value`, must not be `None`")  # noqa: E501
        if value is not None and value < 0:  # noqa: E501
            raise ValueError("Invalid value for `value`, must be a value greater than or equal to `0`")  # noqa: E501

        self._value = value

    @property
    def unit(self) -> CO2Units:
        """Gets the unit of this CO2Forecast.


        :return: The unit of this CO2Forecast.
        :rtype: CO2Units
        """
        return self._unit

    @unit.setter
    def unit(self, unit: CO2Units):
        """Sets the unit of this CO2Forecast.


        :param unit: The unit of this CO2Forecast.
        :type unit: CO2Units
        """
        if unit is None:
            raise ValueError("Invalid value for `unit`, must not be `None`")  # noqa: E501

        self._unit = unit

    @property
    def quality(self) -> float:
        """Gets the quality of this CO2Forecast.


        :return: The quality of this CO2Forecast.
        :rtype: float
        """
        return self._quality

    @quality.setter
    def quality(self, quality: float):
        """Sets the quality of this CO2Forecast.


        :param quality: The quality of this CO2Forecast.
        :type quality: float
        """
        if quality is None:
            raise ValueError("Invalid value for `quality`, must not be `None`")  # noqa: E501
        if quality is not None and quality < 0:  # noqa: E501
            raise ValueError("Invalid value for `quality`, must be a value greater than or equal to `0`")  # noqa: E501

        self._quality = quality

    @property
    def updated_at(self) -> datetime:
        """Gets the updated_at of this CO2Forecast.


        :return: The updated_at of this CO2Forecast.
        :rtype: datetime
        """
        return self._updated_at

    @updated_at.setter
    def updated_at(self, updated_at: datetime):
        """Sets the updated_at of this CO2Forecast.


        :param updated_at: The updated_at of this CO2Forecast.
        :type updated_at: datetime
        """
        if updated_at is None:
            raise ValueError("Invalid value for `updated_at`, must not be `None`")  # noqa: E501

        self._updated_at = updated_at