# statistics_manager_service/openapi/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: vasco.m.campos@inesctec.pt
    name: Vasco Campos
  description: "This is the Statistics Manager Service OpenAPI definition. The Statistics\
    \ Manager Service processes energy data in order to provide the user with useful\
    \ statistics (e.g. energy cost for the day, injection for the month, consumption\
    \ for the year, CO2 footprint). The interactions here represented are based on\
    \ this [Canvas](https://gitlab.inesctec.pt/cpes/european-projects/interconnect/hems/hems-documentation/-/blob/master/Microservices/Statistics-Manager-Service.adoc)"
  title: Statistics Manager Service
  version: 0.1.1
servers:
- url: https://interconnect-dev.inesctec.pt/api/statistics
security:
- {}
- Bearer: []
...
paths:
  /ecosignal-forecast:
    description: Retrieve the CO2 intensity forecast for mix energy of country.
    get:
      operationId: co2_intensity_get
      parameters:
      - explode: false
        in: header
        name: X-Correlation-ID
        required: true
        schema:
          $ref: '#/components/schemas/CorrelationId'
        style: simple
      - explode: false
        in: header
        name: Authorization
        required: false
        schema:
          $ref: '#/components/schemas/Authorization'
        style: simple
      - explode: true
        in: query
        name: forecast_day
        required: true
        schema:
          $ref: '#/components/schemas/Date'
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/CO2Forecast'
                type: array
          description: Success response The timestamp field is the first timestamp
            of the returned values.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "400":
          content:
            application/json:
              examples:
                Timestamp error:
                  value:
                    error: Reason for bad request error
                Missing X-Correlation-ID header:
                  value:
                    error: missing X-Correlation-ID header
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "401":
          content:
            application/json:
              examples:
                Unauthenticated:
                  value:
                    error: error in authorization token
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthenticated or invalid credentials
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "403":
          content:
            application/json:
              examples:
                Forbidden:
                  value:
                    error: you are not authorized to access this information
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden to access the information requested
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "404":
          content:
            application/json:
              examples:
                User Id not found:
                  value:
                    error: user with id "userId" not found
              schema:
                $ref: '#/components/schemas/Error'
          description: User id not found
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
      summary: Retrieve CO2 intensity forecast from sentinel for a single day
      tags:
      - CO2 Forecast
      x-openapi-router-controller: statistics_manager_service.controllers.co2_forecast_controller
...
components:
...
  headers:
    CorrelationId:
      explode: false
      required: true
      schema:
        $ref: '#/components/schemas/CorrelationId'
      style: simple
...
  schemas:
    CorrelationId:
      format: uuid
      type: string
...
    Authorization:
      format: byte
      type: string
    Date:
      example: 2023-01-20
      format: date
      type: string
...
    CO2Units:
      enum:
      - gCO₂eq/kWh
      - KgCO₂eq/KWh
      example: gCO₂eq/kWh
      title: CO2Units
      type: string
    CO2Forecast:
      description: CO2 value related to mostly recent forecast
      example:
        datetime: 2023-07-16T00:00:00Z
        request: 2023-07-15T17:45:02Z
        value: 175.06025393858152
        unit: gCO₂eq/kWh
        quality: 0
        updated_at: 2023-07-15T17:56:01Z
      properties:
        datetime:
          example: 2023-01-20T00:00:00Z
          format: date-time
          title: datetime
          type: string
        request:
          example: 2023-01-20T00:00:00Z
          format: date-time
          title: request
          type: string
        value:
          example: 12.5
          format: float
          minimum: 0
          title: value
          type: number
        unit:
          $ref: '#/components/schemas/CO2Units'
        quality:
          example: 0
          format: int64
          minimum: 0
          title: slot_number
          type: number
        updated_at:
          example: 2023-01-20T00:00:00Z
          format: date-time
          title: updated_at
          type: string
      required:
      - datetime
      - quality
      - request
      - unit
      - updated_at
      - value
      title: CO2Forecast
      type: object
...
    Error:
      description: Error model
      example:
        error: error
      properties:
        error:
          title: error
          type: string
      required:
      - error
      title: Error
      type: object
  securitySchemes:
    Bearer:
      bearerFormat: JWT
      scheme: bearer
      type: http
      x-bearerInfoFunc: statistics_manager_service.controllers.security_controller.info_from_Bearer